-Cập nhật lại code trong form hóa đơn (TranCongMinh) cập nhật lại phần của mình. - Cập nhật file main - Cập nhật lại một vài lỗi nhỏ

+++ QuanLyBanHang/GUI/KhachHang.cs
@@ -1,16 +1,9 @@
-﻿using DAL;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-using BUS;
+﻿using BUS;
+using DAL;
 using DTO;
-using System.Data.SqlClient;
+using System;
+using System.Data;
+using System.Windows.Forms;
 namespace GUI
 {
     public partial class KhachHang : Form
@@ -18,7 +11,7 @@ namespace GUI
         KhachHang_DTO dto = new KhachHang_DTO();
         Lop_DAL dal = new Lop_DAL();
         KhachHang_BUS bus = new KhachHang_BUS();
-        DataTable dtKhachHang,dtTimKiemHang;
+        DataTable dtKhachHang, dtTimKiemHang;
 
         public KhachHang()
         {
@@ -30,7 +23,7 @@ namespace GUI
             dtKhachHang.Clear();
             dtKhachHang = bus.ShowKhachHang();
             dgvKhachHang.DataSource = dtKhachHang;
-            
+
         }
         private void KhachHang_Load(object sender, EventArgs e)
         {
@@ -44,12 +37,12 @@ namespace GUI
             txtDiaChi.ResetText();
             txtSoDienThoai.ResetText();
         }
-        
+
 
         private void btnXem_Click(object sender, EventArgs e)
         {
             LoadData();
-            
+
         }
 
         private void btnSua_Click(object sender, EventArgs e)
@@ -115,13 +108,19 @@ namespace GUI
 
         private void btnXoa_Click(object sender, EventArgs e)
         {
-
-            DialogResult rs = MessageBox.Show("Bạn thực sự muốn xóa \"" + txtHoTen.Text + "\" ra khỏi danh sách?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (rs == DialogResult.Yes)
+            if (txtMaKhachHang.Text == "")
             {
-                bus.DeleteKhachHang(txtMaKhachHang.Text.Trim());
-                MessageBox.Show("Xóa thành công");
-                LoadData();
+                MessageBox.Show("Thất bại, chưa nhập mã khách hàng cần xóa!");
+            }
+            else
+            {
+                DialogResult rs = MessageBox.Show("Bạn thực sự muốn xóa \"" + txtHoTen.Text + "\" ra khỏi danh sách?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                if (rs == DialogResult.Yes)
+                {
+                    bus.DeleteKhachHang(txtMaKhachHang.Text.Trim());
+                    MessageBox.Show("Xóa thành công");
+                    LoadData();
+                }
             }
         }
 
@@ -202,6 +201,11 @@ namespace GUI
             txtHoTen.Text = dgvKhachHang.Rows[r].Cells[1].Value.ToString();
             txtDiaChi.Text = dgvKhachHang.Rows[r].Cells[2].Value.ToString();
             txtSoDienThoai.Text = dgvKhachHang.Rows[r].Cells[3].Value.ToString();
+        }
+
+        private void KhachHang_FormClosed(object sender, FormClosedEventArgs e)
+        {
+
         }
 
         private void btnTim_Click_1(object sender, EventArgs e)
